requirements.txt
fastapi
kafka-python
pydantic
pyspark
redis
requests